# apidocs/openapi.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  version: 1.0.0
  description: |
    OpenAPI specification for the marketplace backend.

    The API uses a consistent envelope format:
    - successful responses: `{ "data": ... }`
    - errors: `{ "error": { "code", "message", "details" } }`
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: /
    description: Relative server URL for reverse proxy setups
...
paths:
...
  /api/v1/products:
    get:
      tags: [Catalog]
      operationId: listProducts
      summary: List products with filtering, sorting, and pagination
      parameters:
        - $ref: '#/components/parameters/SearchQueryParam'
        - $ref: '#/components/parameters/CategoryIDQueryParam'
        - $ref: '#/components/parameters/MinPriceParam'
        - $ref: '#/components/parameters/MaxPriceParam'
        - $ref: '#/components/parameters/InStockParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPageEnvelope'
        '400':
          $ref: '#/components/responses/InvalidInputResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
...
  /api/v1/profile:
    get:
      tags: [Profile]
      operationId: getProfile
      summary: Get the current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
    patch:
      tags: [Profile]
      operationId: updateProfile
      summary: Update profile fields
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
            examples:
              default:
                value:
                  full_name: Ivan Sergeevich Petrov
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelope'
        '400':
          $ref: '#/components/responses/InvalidInputResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
...
  /api/v1/orders:
    get:
      tags: [Orders]
      operationId: listOrders
      summary: List orders for the current user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated order history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPageEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
    post:
      tags: [Orders]
      operationId: checkout
      summary: Checkout the current cart into an order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
            examples:
              default:
                value:
                  place_id: f6d1cb07-b8db-4220-9ab2-bd661e74fb0c
      responses:
        '201':
          description: Order created from the current cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderEnvelope'
        '400':
          $ref: '#/components/responses/CartEmptyOrInvalidInputResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '409':
          $ref: '#/components/responses/ProductConflictResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
...
  /api/v1/seller/dashboard:
    get:
      tags: [Seller]
      operationId: sellerDashboard
      summary: Get seller dashboard summary
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Seller dashboard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellerDashboardEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: mp_access_token
...
  parameters:
    SearchQueryParam:
      name: q
      in: query
      description: Search query for product name, description, brand, slug, SKU, and category.
      required: false
      schema:
        type: string
      example: headphones
    CategoryIDQueryParam:
      name: category_id
      in: query
      description: Filter by category ID.
      required: false
      schema:
        type: string
        format: uuid
      example: 6b997cc0-e76e-4c2f-a0a4-79a9a407f2cc
    MinPriceParam:
      name: min_price
      in: query
      required: false
      schema:
        type: number
        format: double
      example: 100
    MaxPriceParam:
      name: max_price
      in: query
      required: false
      schema:
        type: number
        format: double
      example: 1000
    InStockParam:
      name: in_stock
      in: query
      required: false
      schema:
        type: boolean
      example: true
...
    SortParam:
      name: sort
      in: query
      required: false
      schema:
        type: string
        enum: [new, price_asc, price_desc]
      example: new
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
...
  responses:
    InvalidInputResponse:
      description: Invalid request payload or malformed parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            invalidInput:
              value:
                error:
                  code: invalid_input
                  message: invalid input
...
    UnauthorizedResponse:
      description: Missing or invalid access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            unauthorized:
              value:
                error:
                  code: unauthorized
                  message: unauthorized
...
    ForbiddenResponse:
      description: Authenticated but not allowed to access the resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            forbidden:
              value:
                error:
                  code: forbidden
                  message: forbidden
    NotFoundResponse:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            notFound:
              value:
                error:
                  code: not_found
                  message: resource not found
...
    ProductConflictResponse:
      description: Product cannot be added or checked out because of stock or active state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            insufficientStock:
              value:
                error:
                  code: insufficient_stock
                  message: insufficient stock
            productUnavailable:
              value:
                error:
                  code: product_unavailable
                  message: product unavailable
    CartEmptyOrInvalidInputResponse:
      description: Checkout cannot proceed because the cart is empty or the request is invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            cartEmpty:
              value:
                error:
                  code: cart_empty
                  message: cart is empty
            invalidInput:
              value:
                error:
                  code: invalid_input
                  message: invalid input
...
    InternalErrorResponse:
      description: Unexpected internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            internalError:
              value:
                error:
                  code: internal_error
                  message: internal server error
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: invalid_input
        message:
          type: string
          example: invalid input
        details:
          nullable: true
          description: Optional structured error details.
...
    UpdateProfileRequest:
      type: object
      properties:
        full_name:
          type: string
          maxLength: 120
          example: Ivan Sergeevich Petrov
    User:
      type: object
      required:
        [id, email, role, is_active, is_email_verified, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          example: 7d4c4e98-42ba-4f55-86db-9f3d2772e6c1
        email:
          type: string
          format: email
          example: user@example.com
        full_name:
          type: string
          example: Ivan Petrov
        role:
          type: string
          enum: [customer, seller, admin]
          example: customer
        is_active:
          type: boolean
          example: true
        email_verified_at:
          type: string
          format: date-time
          nullable: true
          example: '2026-03-13T18:15:00Z'
        is_email_verified:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: '2026-03-13T18:10:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2026-03-13T18:15:00Z'
...
    UserEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/User'
...
    Product:
      type: object
      required:
        [id, category_id, name, slug, price, currency, sku, images, stock_qty, is_active, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          example: 3a8b7c10-826a-4046-8b95-93a235e4e0a1
        category_id:
          type: string
          format: uuid
          example: 6b997cc0-e76e-4c2f-a0a4-79a9a407f2cc
        category_name:
          type: string
          example: Электроника и гаджеты
        name:
          type: string
          example: Смартфон
        slug:
          type: string
          example: electronics-gadgets-smartfon
        description:
          type: string
          example: Флагманский смартфон с OLED-экраном, 5G и быстрой зарядкой.
        price:
          type: number
          format: double
          example: 107600
        currency:
          type: string
          example: RUB
        sku:
          type: string
          example: EL-SMART-01
        seller_id:
          type: string
          format: uuid
          nullable: true
          example: 70000000-0000-0000-0000-000000000001
        seller_name:
          type: string
          example: UrbanWave
        seller_slug:
          type: string
          example: urbanwave
        image_url:
          type: string
          example: marketplace-media://product/electronics-gadgets-smartfon/hero
        images:
          type: array
          items:
            type: string
          example:
            - marketplace-media://product/electronics-gadgets-smartfon/hero
            - marketplace-media://product/electronics-gadgets-smartfon/detail
        brand:
          type: string
          example: UrbanWave
        unit:
          type: string
          example: piece
        specs:
          type: object
          additionalProperties: true
          example:
            battery_life: 40 h
            wireless: Bluetooth 5.3
        stock_qty:
          type: integer
          example: 36
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: '2026-03-13T16:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2026-03-13T16:00:00Z'
    ProductPage:
      type: object
      required: [items, page, limit, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 57
...
    ProductPageEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/ProductPage'
...
    CheckoutRequest:
      type: object
      required: [place_id]
      properties:
        place_id:
          type: string
          format: uuid
          example: f6d1cb07-b8db-4220-9ab2-bd661e74fb0c
    OrderItem:
      type: object
      required: [id, order_id, product_id, product_name, sku, unit_price, quantity, line_total, currency, created_at]
      properties:
        id:
          type: string
          format: uuid
          example: ac7f90df-63ba-4e3c-941d-b1f3304b856b
        order_id:
          type: string
          format: uuid
          example: 9ca2b8bc-f925-4df5-9880-34b310ae2f67
        product_id:
          type: string
          format: uuid
          example: 3a8b7c10-826a-4046-8b95-93a235e4e0a1
        product_name:
          type: string
          example: Смартфон
        sku:
          type: string
          example: EL-SMART-01
        seller_id:
          type: string
          format: uuid
          nullable: true
          example: 70000000-0000-0000-0000-000000000001
        seller_name:
          type: string
          example: UrbanWave
        unit_price:
          type: number
          format: double
          example: 11990
        quantity:
          type: integer
          example: 1
        line_total:
          type: number
          format: double
          example: 11990
        currency:
          type: string
          example: RUB
        created_at:
          type: string
          format: date-time
          example: '2026-03-13T18:25:00Z'
    Order:
      type: object
      required:
        [id, user_id, place_id, place_title, address_text, status, currency, total_amount, created_at, updated_at, items]
      properties:
        id:
          type: string
          format: uuid
          example: 9ca2b8bc-f925-4df5-9880-34b310ae2f67
        user_id:
          type: string
          format: uuid
          example: 7d4c4e98-42ba-4f55-86db-9f3d2772e6c1
        place_id:
          type: string
          format: uuid
          example: f6d1cb07-b8db-4220-9ab2-bd661e74fb0c
        place_title:
          type: string
          example: Home
        address_text:
          type: string
          example: Moscow, Tverskaya 1
        lat:
          type: number
          format: double
          nullable: true
          example: 55.7558
        lon:
          type: number
          format: double
          nullable: true
          example: 37.6173
        status:
          type: string
          enum: [pending, cancelled]
          example: pending
        currency:
          type: string
          example: RUB
        total_amount:
          type: number
          format: double
          example: 991
        created_at:
          type: string
          format: date-time
          example: '2026-03-13T18:25:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2026-03-13T18:25:00Z'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    OrderPage:
      type: object
      required: [items, page, limit, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 4
    OrderEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/Order'
    OrderPageEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/OrderPage'
    SellerProfile:
      type: object
      required: [user_id, store_name, store_slug, status, created_at, updated_at]
      properties:
        user_id:
          type: string
          format: uuid
          example: 70000000-0000-0000-0000-000000000001
        store_name:
          type: string
          example: UrbanWave
        store_slug:
          type: string
          example: urbanwave
        legal_name:
          type: string
          example: UrbanWave Trading LLC
        description:
          type: string
          example: Audio and lifestyle accessories with fast dispatch and city pickup.
        logo_url:
          type: string
          example: https://cdn.example.com/stores/urbanwave-logo.jpg
        banner_url:
          type: string
          example: https://cdn.example.com/stores/urbanwave-banner.jpg
        support_email:
          type: string
          format: email
          example: hello@urbanwave.example
        support_phone:
          type: string
          example: +7 999 555-10-10
        city:
          type: string
          example: Moscow
        status:
          type: string
          enum: [pending, active, paused]
          example: active
        created_at:
          type: string
          format: date-time
          example: '2026-03-18T09:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2026-03-19T09:00:00Z'
...
    SellerMetrics:
      type: object
      required: [products_total, active_products, hidden_products, low_stock_products, orders_total, units_sold, gross_revenue]
      properties:
        products_total:
          type: integer
          example: 18
        active_products:
          type: integer
          example: 14
        hidden_products:
          type: integer
          example: 4
        low_stock_products:
          type: integer
          example: 3
        orders_total:
          type: integer
          example: 27
        units_sold:
          type: integer
          example: 63
        gross_revenue:
          type: number
          format: double
          example: 184950
    SellerOrderSummary:
      type: object
      required: [order_id, status, currency, created_at, place_title, items_count, gross_revenue]
      properties:
        order_id:
          type: string
          format: uuid
          example: 9ca2b8bc-f925-4df5-9880-34b310ae2f67
        status:
          type: string
          enum: [pending, cancelled]
          example: pending
        currency:
          type: string
          example: RUB
        created_at:
          type: string
          format: date-time
          example: '2026-03-19T10:15:00Z'
        place_title:
          type: string
          example: Home
        items_count:
          type: integer
          example: 2
        gross_revenue:
          type: number
          format: double
          example: 21480
...
    SellerDashboard:
      type: object
      required: [profile, metrics, recent_products, low_stock, recent_orders]
      properties:
        profile:
          $ref: '#/components/schemas/SellerProfile'
        metrics:
          $ref: '#/components/schemas/SellerMetrics'
        recent_products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        low_stock:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        recent_orders:
          type: array
          items:
            $ref: '#/components/schemas/SellerOrderSummary'
    SellerDashboardEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/SellerDashboard'
...